Support specification of page size in the UI

+++ webprotege-gwt-ui-client/src/main/java/edu/stanford/bmir/protege/web/client/form/RepeatabilityView.java
@@ -17,7 +17,7 @@ public interface RepeatabilityView extends IsWidget {
     @Nonnull
     Repeatability getRepeatability();
 
-    long getPageSize();
+    int getPageSize();
 
-    void setPageSize(long pageSize);
+    void setPageSize(int pageSize);
 }
